FIT: added explicit type conversion where EEventDataBit is used

+++ DataFormats/Detectors/FIT/FDD/include/DataFormatsFDD/RecPoint.h
@@ -49,11 +49,11 @@ struct ChannelDataFloat {
 
   static void setFlag(fit::EEventDataBit bitFlag, int& adcId)
   {
-    adcId = uint8_t(adcId) | 1u << bitFlag;
+    adcId = uint8_t(adcId) | 1u << uint8_t(bitFlag);
   }
   static void clearFlag(fit::EEventDataBit bitFlag, int& adcId)
   {
-    adcId = uint8_t(adcId) & ~(1u << bitFlag);
+    adcId = uint8_t(adcId) & ~(1u << uint8_t(bitFlag));
   }
   void setFlag(int flag)
   {
@@ -61,11 +61,11 @@ struct ChannelDataFloat {
   }
   void setFlag(fit::EEventDataBit bitFlag, bool value)
   {
-    adcId = uint8_t(adcId) | uint8_t(value) << bitFlag;
+    adcId = uint8_t(adcId) | uint8_t(value) << uint8_t(bitFlag);
   }
   bool getFlag(fit::EEventDataBit bitFlag) const
   {
-    return bool(uint8_t(adcId) & (1u << bitFlag));
+    return bool(uint8_t(adcId) & (1u << uint8_t(bitFlag)));
   }
   bool areAllFlagsGood() const
   {
